Move aria states handler into states object
