+++ lib/config.js
@@ -79,4 +79,6 @@ function isLegacy()
     return false;
 }
 
-module.exports = getTophatConfig;
+var generatedConfig = getTophatConfig();
+
+module.exports = generatedConfig;
