+++ lib/events/index.js
@@ -33,7 +33,7 @@ function attachTophatEvents( document, tophatElement ) {
 // Hande the window resizing
 function resizeHandler(e) {
     console.log("window resize handler");
-    utils.setAriaStates();
+    this.state.setAriaStates();
 }
 
 function trackingHandler( ev ) {
