+++ lib/events/states.js
@@ -1,7 +1,9 @@
-var utils = require('../utils/dom');
-var evidenceStateClassname = 'menu-evidence-open';
-var profileStateClassname = 'menu-profile-open';
-var mobileStateClassname = 'menu-mobile-open';
+var utils = require('../utils/dom'),
+    config = require('../config');
+
+var evidenceStateClassname = 'menu-evidence-open',
+    profileStateClassname = 'menu-profile-open',
+    mobileStateClassname = 'menu-mobile-open';
 
 function TophatStates( el ) {
     this.element = el;
@@ -15,6 +17,8 @@ function TophatStates( el ) {
     this.profileBtn = document.getElementById("menu-profile");
 
     el.state = this;
+
+    this.setAriaStates();
 }
 
 function getStateFromClassname( classname ) {
@@ -99,6 +103,38 @@ TophatStates.prototype = {
         this.data.mobile = false;
         this.data.evidence = false;
         this.updateState();
+    }
+    
+    // Sets the correct state of aria attributes depending on device.
+    , setAriaStates: function(){
+        var mobile = document.body.clientWidth < 767;
+        var mobileMenu = document.getElementById("menu-mobile");
+        var mobileEvidenceMenu = document.getElementById("menu-evidence");
+        var mobileDropdown = document.getElementById("nice-evidence");
+        var mainMenu = document.getElementById("main-menu");
+
+        function setMobileState(bool){
+            var mobileState = bool;
+            mobileMenu.setAttribute("aria-hidden",!bool);
+            mainMenu.setAttribute("aria-hidden",bool);
+            mobileDropdown.setAttribute("aria-hidden",bool);
+            mobileEvidenceMenu.setAttribute("aria-expanded",bool);
+        }
+
+        function setEvidenceServiceStates(){
+            var service = config.evidence;
+            if(service !== undefined)
+                mobileEvidenceMenu.setAttribute("aria-expanded","true");
+        }
+        
+        if(mobile)
+             setMobileState(true); 
+
+        else
+            setMobileState(false);
+
+        setEvidenceServiceStates();
+
     }
 
 };
