+++ lib/tophat.js
@@ -5,7 +5,7 @@ require('./tophat.css');
 var utils = require('./utils/dom');
 var body = document.body;
 
-var config = require('./config')();
+var config = require('./config');
 
 // find or create the tophat element
 var tophatElement = getTophatElement( 'nice-tophat', config );
@@ -69,6 +69,3 @@ function composeTophat( el, services, evidenceResources, globalMenu, config ) {
     // attach the tophat to the top of the body
     utils.prependElement( tophatElement, body );
 }
-
-//Sets the correct state of aria attributes depending on device.
-utils.setAriaStates();
