+++ lib/utils/dom.js
@@ -174,35 +174,4 @@ utils.appendElement = function( element, parent ) {
     parent.appendChild( element );
 };
 
-
-utils.setAriaStates = function(){
-    var mobile = document.body.clientWidth < 767;
-    var mobileMenu = document.getElementById("menu-mobile");
-    var mobileEvidenceMenu = document.getElementById("menu-evidence");
-    var mobileDropdown = document.getElementById("nice-evidence");
-    var mainMenu = document.getElementById("main-menu");
-
-    function setMobileState(bool){
-        var mobileState = bool;
-        mobileMenu.setAttribute("aria-hidden",!bool);
-        mainMenu.setAttribute("aria-hidden",bool);
-        mobileDropdown.setAttribute("aria-hidden",bool);
-        mobileEvidenceMenu.setAttribute("aria-expanded",bool);
-    }
-
-    function setEvidenceServiceStates(){
-        var service = config().evidence;
-        if(service !== undefined)
-            mobileEvidenceMenu.setAttribute("aria-expanded","true");
-    }
-    
-    if(mobile)
-         setMobileState(true); 
-
-    else
-        setMobileState(false);
-
-    setEvidenceServiceStates();
-
-};
 module.exports = utils;
